# [ V0 ][ sample ] improve sample performance when using  guide decoding