--- a/vllm/model_executor/models/interfaces.py
+++ b/vllm/model_executor/models/interfaces.py
@@ -12,6 +12,9 @@ from vllm.model_executor.layers.quantization.base_config import (
     QuantizationConfig)
 from vllm.utils import supports_kw
 
+from .. import SamplingMetadata
+from ..layers.logits_processor import _apply_logits_processors, accept_grammar
+from ..layers.sampler import SamplerOutput
 from .interfaces_base import is_pooling_model
 
 if TYPE_CHECKING:
@@ -219,6 +222,55 @@ class SupportsPP(Protocol):
         Return :class:`IntermediateTensors` only for the last PP rank.
         """
         ...
+
+
+class SupportsSampleV2:
+
+    def compute_logits_v2(
+        self,
+        hidden_states: torch.Tensor,
+        sampling_metadata: SamplingMetadata,
+    ) -> Optional[torch.Tensor]:
+        logits = self.logits_processor(self.lm_head,
+                                       hidden_states,
+                                       sampling_metadata,
+                                       skip_grammar=True)
+        return logits
+
+    def samplev2(
+        self,
+        logits: torch.Tensor,
+        sampling_metadata: SamplingMetadata,
+    ) -> Optional[SamplerOutput]:
+        # compute logits
+        next_tokens: SamplerOutput = self.sampler(logits, sampling_metadata)
+
+        # check if  the sampled tokens fit the grammars
+        tks = torch.tensor(
+            [o.samples[0].output_token for o in next_tokens.outputs])
+        accepted = accept_grammar(tks, sampling_metadata)
+        need_resample = torch.logical_not(accepted)
+        if accepted.all():
+            return next_tokens
+        # resample
+        # if the token is not valid, sample again.
+        # but first apply the grammar bitmask
+        # only apply logits processor when need_resample
+        logits = _apply_logits_processors(logits, sampling_metadata,
+                                          need_resample, False)
+        new_next_tokens: SamplerOutput = self.sampler(logits,
+                                                      sampling_metadata)
+
+        for i, replace in enumerate(need_resample.tolist()):
+            if replace:
+                next_tokens.outputs[i] = new_next_tokens.outputs[i]
+
+        tks = torch.tensor(
+            [o.samples[0].output_token for o in next_tokens.outputs])
+        # matcher only accept next token when first round is not accepted.
+        accepted = accept_grammar(tks, sampling_metadata, need_resample)
+        assert accepted.all()
+        return next_tokens
 
 
 # We can't use runtime_checkable with ClassVar for issubclass checks