@russellb 2025-03-21
> > Sorry, can you merge from latest main to fix the pre-commit failures?
> 
> ok. l have fixed the pre-commit faulures

All commit messages also need `Signed-off-by` headers to make the DCO check pass.

About the change -- the results are impressive, but I'm a bit concerned about what we'd do with this for V1. As written, this will only work with the feature for V0, but we're trying to focus our enhancements on V1 as much as possible.

Have you compared this to V1? That would be good as another column in your comparison. In other words, does V0 + this enhancement beat V1 with structured output in use?  Or do the other enhancements in V1 already make V1 faster without this optimization in place?