@cjsdurj 2025-03-22
> > @cjsdurj did you mean to close this?
> > sorry. I have make a mistake last night when address DCO check.
> 
> * this feature also work on V1. and I am working on it.
> * I have  test  v1  engine  , the gbnf decode performance is  as  slow as v0 .
> 
> because it changes too many py files , currently it only works on V0 and qwen llama model as a preview feature.

 if this works ok.  In subsequent submissions,  I will  Implement  this feature in V1 engine  for more models.