@russellb 2025-03-22
> > > @cjsdurj did you mean to close this?
> > > sorry. I have make a mistake last night when address DCO check.
> > 
> > 
> > 
> > * this feature also work on V1. and I am working on it.
> > * I have  test  v1  engine  , the gbnf decode performance is  as  slow as v0 .
> > 
> > because it changes too many py files , currently it only works on V0 and qwen llama model as a preview feature.
> 
> if this works ok. In subsequent submissions, I will Implement this feature in V1 engine for more models.

It will not work with V1 by design right now. In V1, advancing the grammar's FSM and applying the bitmask are in separate processes.

I'd like to see more performance numbers, in particular with large batches of requests and not just a single request. I'll do some benchmarking at some point if you don't have the hardware for it (I'll want to see some H100 results).