@lk-chen 2025-04-29
https://github.com/vllm-project/vllm/pull/17084 removed sampler from model, this PR needs rebase.
Let me see if I can help