@lk-chen 2025-04-30
Hi @cjsdurj , may I ask how to produce before throughput of 2tk/s and after throughput of 136 tk/s ?

I'm using https://github.com/lk-chen/vllm/pull/2 on L40S, forcing **vLLM v0, model=Qwen/Qwen2.5-1.5B-Instruct, async mode**, and got

|num_prompts| output trpt. before this pr (tk/s) | output trpt. after this pr|
|----------|----------|----------|
|1      |  12.94 | 12.95 |
|100 | 69.41 | 79.33 |